Stub out references to e.clone() until later.
Unfortunately the generic `failure::Error` type is not `Clone` because
of its internals: https://github.com/rust-lang-nursery/failure/issues/148

The fix is to either switch to `Rc<Error>` or `Arc<Error>`, or to use a
custom `Fail` type.  In this case using a custom `Fail` might be best
but this commit just stubs out references to e.clone() until then.