Better stream ownership in PeerServer, rm stubs
We want the PeerServer to be generic over the kind of stream that it
has, so that the PeerConnector can use stream combinators to add hooks
into the streams that it constructs (e.g., to fire last-seen
timestamps).

A more natural design would be for the PeerServer struct to own both tx
and rx halves of the message stream, but this complicates the borrow
checker (since the main event loop needs to hold a future that controls
rx, it cannot allow something else to have mutable access to rx, so if
the rx end is part of self, it cannot make &mut self calls).