consensus: Call AddBlock after all checks pass
We don't want to call the state's AddBlock until we know the block is
valid. This avoids subtle bugs if the state is modified in the call().

(in_memory currently modifies the state in the call(), on_disk modifies
the state in the async block.)

Also moves verification into the async block.

Part of #429.