consensus: Conditionally call the state's poll_ready
We avoid calling the state's poll_ready() in the checkpoint verifier's
poll_ready. This change allows us to conditionally make AddBlock
requests, via a buffered state instance.

(Tower::Buffer expects a 1:1 relationship between poll_ready() and
call().)

This change breaks backpressure propagation between the state and
verifier users. That's ok, because we don't expect state changes to be
slow enough to exert backpressure on the network. (Verifiers can still
exert backpressure on the network, if verification is slow.)

Part of #429.